# locaFastApi/swagger-config.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Api LocaFast",
    "description": "Documentação da API LocaFast",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
...
    "/cliente": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "summary": "Listar todos os clientes",
        "description": "Retorna uma lista de todos os clientes cadastrados.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes",
            "content": {
              "application/json": {
                "example": {
                  "numeroCnh": "123456789",
                  "validadeCnh": "2023-12-31",
                  "estadoEmissor": "SP",
                  "nome": "Exemplo Cliente",
                  "cpf": "123.456.789-01",
                  "rg": "12345678",
                  "telefone": "(12)3456-7890",
                  "email": "cliente@example.com",
                  "dataNascimento": "1990-01-01",
                  "status": "Ativo",
                  "genero": "Feminino"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro ao buscar clientes."
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clientes"
        ],
        "summary": "Cadastrar um novo cliente",
        "description": "Cadastra um novo cliente com as informações fornecidas.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "numeroCnh": "123456789",
                "validadeCnh": "2023-12-31",
                "estadoEmissor": "SP",
                "nome": "Exemplo Cliente",
                "cpf": "123.456.789-01",
                "rg": "12345678",
                "telefone": "(12)3456-7890",
                "email": "cliente@example.com",
                "dataNascimento": "1990-01-01",
                "status": "Ativo",
                "genero": "Feminino"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sucesso ao cadastrar cliente",
            "content": {
              "application/json": {
                "example": {
                  "message": "Cliente cadastrado com sucesso!"
                }
              }
            }
          },
          "400": {
            "description": "Falha na verificação de uma regra de negócio",
            "content": {
              "application/json": {
                "example": {
                  "message 1": "Todos os campos devem ser preenchidos.",
                  "message 2": "número da CNH inválido.",
                  "message 3": "CPF inválido.",
                  "message 4": "RG inválido.",
                  "message 5": "O telefone inserido é inválido.",
                  "message 6": "O email inserido é inválido."
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro ao cadastrar cliente."
                }
              }
            }
          }
        }
      }
    },
    "/cliente/{id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "summary": "Obter cliente por ID",
        "description": "Retorna um cliente com o ID fornecido.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do cliente",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado",
            "content": {
              "application/json": {
                "example": {
                  "id": "65ede1b972f9e877b79e0ef9"
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado",
            "content": {
              "application/json": {
                "example": {
                  "message": "Cliente não encontrado."
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro ao buscar cliente por ID."
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Clientes"
        ],
        "summary": "Atualizar cliente por ID",
        "description": "Atualiza um cliente com o ID fornecido.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do cliente",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "numeroCnh": "123456789",
                "validadeCnh": "2023-12-31",
                "estadoEmissor": "SP",
                "nome": "Exemplo Cliente",
                "cpf": "123.456.789-01",
                "rg": "12345678",
                "telefone": "(12)3456-7890",
                "email": "cliente@example.com",
                "dataNascimento": "1990-01-01",
                "status": "Ativo",
                "genero": "Feminino"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cliente atualizado com sucesso",
            "content": {
              "application/json": {
                "example": {
                  "message": "Cliente atualizado com sucesso!"
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado",
            "content": {
              "application/json": {
                "example": {
                  "message": "Cliente não existe."
                }
              }
            }
          },
          "400": {
            "description": "Falha na verificação de uma regra de negócio",
            "content": {
              "application/json": {
                "example": {
                  "message 1": "Todos os campos devem ser preenchidos.",
                  "message 2": "número da CNH inválido.",
                  "message 3": "CPF inválido.",
                  "message 4": "RG inválido.",
                  "message 5": "O telefone inserido é inválido.",
                  "message 6": "O email inserido é inválido."
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro ao atualizar cliente."
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clientes"
        ],
        "summary": "Remover cliente por ID",
        "description": "Remove um cliente com o ID fornecido.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do cliente",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente removido com sucesso",
            "content": {
              "application/json": {
                "example": {
                  "message": "Cliente removido com sucesso!"
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado",
            "content": {
              "application/json": {
                "example": {
                  "message": "Cliente não existe."
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro ao remover cliente."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "description": "Autenticação JWT",
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}